move GMath
